Refactor to setup for generic generators
