+++ src/commands/generate/generators/auth/handlers/allow.js
@@ -1,0 +1,11 @@
+module.exports = function () {
+  return `// For production allow list add users below where keys match unique property of OAuth user
+export default {
+    appAccounts: {
+        // janedoe: {
+        //   role: 'member',
+        //   name: 'Jane Doe'
+        // }
+    }
+}`
+}
