+++ src/commands/generate/generators/auth/handlers/html-page-example.js
@@ -1,0 +1,6 @@
+module.exports = function () {
+  return `export default function({html,state}) {
+    const user = state?.store?.account
+    return html\`<p> Hello \${ user?.name }</p>\`
+  }`
+}
