+++ src/commands/generate/generators/auth/handlers/json-api-example.js
@@ -1,0 +1,18 @@
+module.exports = function () {
+  return `import arcOauth from 'arc-plugin-oauth'
+const checkAuth = arcOauth.checkAuth
+
+export async function get(req) {
+    const authenticated = checkAuth(req)
+    if (authenticated) {
+      return {
+        json: { data: ['fred', 'joe', 'mary'] }
+      }
+    } else {
+      return {
+        statusCode:401,
+        json:{error:"not authorized"}
+      }
+    }
+}`
+}
