+++ src/commands/generate/generators/auth/manifest.js
@@ -1,0 +1,20 @@
+const manifest = {
+  routes: [],
+  sourceFiles: [
+    // Utils
+    { src: 'handlers/allow.js', target: 'models/auth/allow.mjs' },
+    { src: 'handlers/mock-allow.js', target: 'models/auth/mock-allow.mjs' },
+    // API
+    { src: 'handlers/html-api-example.js', target: 'app/api/auth/html-example.mjs' },
+    { src: 'handlers/json-api-example.js', target: 'app/api/auth/json-example.mjs' },
+    // Pages
+    { src: 'handlers/html-page-example.js', target: 'app/pages/auth/html-example.mjs' },
+  ],
+  elements: [
+  ],
+  dependencies: [
+    'arc-plugin-oauth'
+  ]
+}
+
+module.exports = manifest
